+++ transformers/examples/text-classification/run_glue_synthetic.py
@@ -171,6 +171,7 @@ class DataTrainingArguments:
             self.task_name = self.task_name.lower()
             if self.task_name not in task_to_keys.keys():
                 raise ValueError("Unknown task, you should pick one in " + ",".join(task_to_keys.keys()))
+        # NOTE: Modifying this. We don't need a train_file for evaluation
         elif self.train_file is None or self.validation_file is None:
             raise ValueError("Need either a GLUE task or a training/validation file.")
         else: