+++ transformers/examples/token-classification/run_ner_synthetic.py
@@ -316,6 +316,7 @@ def main():
         model_args.tokenizer_name if model_args.tokenizer_name else model_args.model_name_or_path,
         cache_dir=model_args.cache_dir,
         use_fast=True,
+        add_prefix_space=True,
     )
     model = AutoModelForTokenClassification.from_pretrained(
         model_args.model_name_or_path,